requirements.txt
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-WTF==0.14.2
gunicorn==19.9.0
iso-639==0.4.5
langdetect==1.0.7
whitenoise==4.1.4
